+++ examples/microprofile-graphql/src/main/java/com/blazebit/persistence/examples/microprofile/graphql/config/CriteriaBuilderFactoryProducer.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2014 - 2021 Blazebit.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.blazebit.persistence.examples.microprofile.graphql.config;
+
+import com.blazebit.persistence.Criteria;
+import com.blazebit.persistence.CriteriaBuilderFactory;
+import com.blazebit.persistence.spi.CriteriaBuilderConfiguration;
+
+import javax.annotation.PostConstruct;
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.context.Initialized;
+import javax.enterprise.event.Observes;
+import javax.enterprise.inject.Produces;
+import javax.inject.Inject;
+import javax.persistence.EntityManagerFactory;
+
+/**
+ * @author Christian Beikov
+ * @since 1.6.2
+ */
+@ApplicationScoped
+public class CriteriaBuilderFactoryProducer {
+
+    // inject your entity manager factory
+    @Inject
+    private EntityManagerFactory entityManagerFactory;
+
+    private CriteriaBuilderFactory criteriaBuilderFactory;
+
+    @PostConstruct
+    public void init() {
+        CriteriaBuilderConfiguration config = Criteria.getDefault();
+        // do some configuration
+        this.criteriaBuilderFactory = config.createCriteriaBuilderFactory(entityManagerFactory);
+    }
+
+    void onStart(@Observes @Initialized(ApplicationScoped.class) Object ev) {
+        // Empty observer to trigger eager bean initialization
+    }
+
+    @Produces
+    @ApplicationScoped
+    public CriteriaBuilderFactory createCriteriaBuilderFactory() {
+        return criteriaBuilderFactory;
+    }
+
+}