+++ examples/microprofile-graphql/src/main/java/com/blazebit/persistence/examples/microprofile/graphql/config/EntityViewManagerProducer.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2014 - 2021 Blazebit.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.blazebit.persistence.examples.microprofile.graphql.config;
+
+import com.blazebit.persistence.CriteriaBuilderFactory;
+import com.blazebit.persistence.view.EntityViewManager;
+import com.blazebit.persistence.view.spi.EntityViewConfiguration;
+
+import javax.annotation.PostConstruct;
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.context.Initialized;
+import javax.enterprise.event.Observes;
+import javax.enterprise.inject.Produces;
+import javax.inject.Inject;
+
+/**
+ * @author Christian Beikov
+ * @since 1.6.2
+ */
+@ApplicationScoped
+public class EntityViewManagerProducer {
+
+    // inject the configuration provided by the cdi integration
+    @Inject
+    private EntityViewConfiguration config;
+
+    // inject the criteria builder factory which will be used along with the entity view manager
+    @Inject
+    private CriteriaBuilderFactory criteriaBuilderFactory;
+
+    private volatile EntityViewManager evm;
+
+    @PostConstruct
+    public void init() {
+        // do some configuration
+        evm = config.createEntityViewManager(criteriaBuilderFactory);
+    }
+
+    void onStart(@Observes @Initialized(ApplicationScoped.class) Object ev) {
+        // Empty observer to trigger eager bean initialization
+    }
+
+    @Produces
+    @ApplicationScoped
+    public EntityViewManager createEntityViewManager() {
+        return evm;
+    }
+}